iOS 8 changes along with Xcode 6.2  and 64bit warning fixes. URL update to new server

+++ Benthos.m
@@ -389,7 +389,7 @@ static sqlite3_stmt *deleteBondSQLStatement = nil;
     sqlite3_bind_double(updateModelSQLStatement, 7, (double)coord.longitude);
     sqlite3_bind_double(updateModelSQLStatement, 8, (double)0);
 
-	sqlite3_bind_int(updateModelSQLStatement, 9, databaseKey);
+	sqlite3_bind_int(updateModelSQLStatement, 9, (int)databaseKey);
 
 	// Execute the query.
 	int success = sqlite3_step(updateModelSQLStatement);
@@ -411,7 +411,7 @@ static sqlite3_stmt *deleteBondSQLStatement = nil;
         }
     }
 	// Bind the query variables.
-	sqlite3_bind_int(insertMetadataSQLStatement, 1, databaseKey);
+	sqlite3_bind_int(insertMetadataSQLStatement, 1, (int)databaseKey);
 	sqlite3_bind_int(insertMetadataSQLStatement, 2, metadataType);
 	sqlite3_bind_text(insertMetadataSQLStatement, 3, [[metadata stringByReplacingOccurrencesOfString:@"'" withString:@"''"] UTF8String], -1, SQLITE_TRANSIENT);
     int success = sqlite3_step(insertMetadataSQLStatement);
@@ -433,9 +433,9 @@ static sqlite3_stmt *deleteBondSQLStatement = nil;
     }
 	// Bind the query variables.
 	sqlite3_clear_bindings(insertAtomSQLStatement);
-	sqlite3_bind_int(insertAtomSQLStatement, 1, databaseKey);
+	sqlite3_bind_int(insertAtomSQLStatement, 1, (int)databaseKey);
 	sqlite3_bind_int(insertAtomSQLStatement, 2, residueKey);
-	sqlite3_bind_int(insertAtomSQLStatement, 3, structureNumber);
+	sqlite3_bind_int(insertAtomSQLStatement, 3, (int)structureNumber);
 	sqlite3_bind_int(insertAtomSQLStatement, 4, atomType);
 	sqlite3_bind_double(insertAtomSQLStatement, 5, (double)newPoint.x);
 	sqlite3_bind_double(insertAtomSQLStatement, 6, (double)newPoint.y);
@@ -483,9 +483,9 @@ static sqlite3_stmt *deleteBondSQLStatement = nil;
     }
 	// Bind the query variables.
 	sqlite3_clear_bindings(insertBondSQLStatement);
-	sqlite3_bind_int(insertBondSQLStatement, 1, databaseKey);
-	sqlite3_bind_int(insertBondSQLStatement, 2, residueKey);
-	sqlite3_bind_int(insertBondSQLStatement, 3, structureNumber);
+	sqlite3_bind_int(insertBondSQLStatement, 1, (int)databaseKey);
+	sqlite3_bind_int(insertBondSQLStatement, 2, (int)residueKey);
+	sqlite3_bind_int(insertBondSQLStatement, 3, (int)structureNumber);
 	sqlite3_bind_int(insertBondSQLStatement, 4, bondType);
 	sqlite3_bind_double(insertBondSQLStatement, 5, (double)startPoint.x);
 	sqlite3_bind_double(insertBondSQLStatement, 6, (double)startPoint.y);
@@ -519,7 +519,7 @@ static sqlite3_stmt *deleteBondSQLStatement = nil;
 	}
 	
 	// Bind the query variables.
-	sqlite3_bind_int(retrieveMetadataSQLStatement, 1, databaseKey);
+	sqlite3_bind_int(retrieveMetadataSQLStatement, 1, (int)databaseKey);
 
 	while (sqlite3_step(retrieveMetadataSQLStatement) == SQLITE_ROW) 
 	{
@@ -577,7 +577,7 @@ static sqlite3_stmt *deleteBondSQLStatement = nil;
 		if (sqlite3_prepare_v2(database, sql, -1, &deleteModelSQLStatement, NULL) != SQLITE_OK) 
 			NSAssert1(0, NSLocalizedStringFromTable(@"Error: failed to prepare statement with message '%s'.", @"Localized", nil), sqlite3_errmsg(database));
 	}
-	sqlite3_bind_int(deleteModelSQLStatement, 1, databaseKey);
+	sqlite3_bind_int(deleteModelSQLStatement, 1, (int)databaseKey);
 	int success = sqlite3_step(deleteModelSQLStatement);
 	sqlite3_reset(deleteModelSQLStatement);
 	if (success != SQLITE_DONE) 
@@ -590,7 +590,7 @@ static sqlite3_stmt *deleteBondSQLStatement = nil;
 		if (sqlite3_prepare_v2(database, sql, -1, &deleteMetadataSQLStatement, NULL) != SQLITE_OK) 
 			NSAssert1(0, NSLocalizedStringFromTable(@"Error: failed to prepare statement with message '%s'.", @"Localized", nil), sqlite3_errmsg(database));
 	}
-	sqlite3_bind_int(deleteMetadataSQLStatement, 1, databaseKey);
+	sqlite3_bind_int(deleteMetadataSQLStatement, 1, (int)databaseKey);
 	success = sqlite3_step(deleteMetadataSQLStatement);
 	sqlite3_reset(deleteMetadataSQLStatement);
 	if (success != SQLITE_DONE) 
@@ -607,7 +607,7 @@ static sqlite3_stmt *deleteBondSQLStatement = nil;
     
 	if (sqlite3_prepare_v2(database, sql, -1, &atomCountingStatement, NULL) == SQLITE_OK) 
 	{
-        sqlite3_bind_int(atomCountingStatement, 1, databaseKey);
+        sqlite3_bind_int(atomCountingStatement, 1, (int)databaseKey);
         sqlite3_bind_int(atomCountingStatement, 2, numberOfStructureBeingDisplayed);
         
         if (sqlite3_step(atomCountingStatement) == SQLITE_ROW)
@@ -632,7 +632,7 @@ static sqlite3_stmt *deleteBondSQLStatement = nil;
     
 	if (sqlite3_prepare_v2(database, sql, -1, &bondCountingStatement, NULL) == SQLITE_OK) 
 	{
-        sqlite3_bind_int(bondCountingStatement, 1, databaseKey);
+        sqlite3_bind_int(bondCountingStatement, 1, (int)databaseKey);
         sqlite3_bind_int(bondCountingStatement, 2, numberOfStructureBeingDisplayed);
         
         if (sqlite3_step(bondCountingStatement) == SQLITE_ROW)
@@ -778,7 +778,7 @@ static sqlite3_stmt *deleteBondSQLStatement = nil;
 	}
 	
 	// Bind the query variables.
-	sqlite3_bind_int(retrieveAtomSQLStatement, 1, databaseKey);
+	sqlite3_bind_int(retrieveAtomSQLStatement, 1,(int) databaseKey);
 	sqlite3_bind_int(retrieveAtomSQLStatement, 2, numberOfStructureBeingDisplayed);
 	
 	while ((sqlite3_step(retrieveAtomSQLStatement) == SQLITE_ROW) && !isRenderingCancelled)
